Added edit page

+++ src/controllers/movieController.js
@@ -61,9 +61,21 @@ movieController.post('/:movieId/attach', isAuth, async(req, res) => {
 
 movieController.get('/:movieId/delete', isAuth, async(req, res) => {
     const movieId = req.params.movieId;
+    const movie = movieService.getOne(movieId);
+
+    if (!movie.ownerId?.equals(req.user.id)) {
+        return res.redirect('/');
+    }
 
     await movieService.delete(movieId);
     res.redirect('/');
+})
+
+movieController.get('/:movieId/edit', isAuth, async(req, res) => {
+    const movieId = req.params.movieId;
+    const movie = await movieService.getOne(movieId);
+
+    res.render('movies/edit', { movie, pageTitle: "Edit" })
 })
 
 export default movieController;
